+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @placement/vue
 
+## 0.2.3
+
+### Patch Changes
+
+- fix point bugs, loosen props
+- Updated dependencies
+  - placement@0.3.3
+
+## 0.2.2
+
+### Patch Changes
+
+- fix some bugs
+- Updated dependencies
+  - placement@0.3.2
+
 ## 0.2.1
 
 ### Patch Changes